requirements.txt
streamlit
langchain
chromadb
openai
tiktoken
pandas
altair
scikit-learn
pyspark
requests
mistralai
python-dotenv
